Move core classes unto the Core namespace
This moves all core classes under the Core namespace.  This way, each namespace, such as User, Plugin, etc are not each their own core plugins.  This should make it easier to tell if a class is a core plugin, and it will consolidate all 'core' functionality and hopefully make the project cleaner.