Don't fire off a db query on __construct
We can't be sure that a proper db connection has been established yet.  For example, the db tests run a pre-setup commend, but if we connect before we get to that point, we could be connecting to the production db.  I will have to do more to prevent this from happening.