Improve auto-queue list
The old query getting the latest scan that was not running, queued, or created.  That means that if there was a scan queued for the site, it would try to queue it again.  The new query gets the latest scan for each site and then checks its status.  It also grabs sites that have never been scanned.