Scan the RFC3986 version of the URI
Scanning a uri such as `http://example.org/example?test test test` will result in a 400 error via curl, but a browser will auto-encode it to `http://example.org/example?test%20test%20test`
This can cause confusion, because the page that a developer sees is different than what sitemaster scanned.

To work around this we can mimic the behavior of the browser by scanning the RFC3986 version of the uri.  This ensures:

Note that the w3c validator will report an error for links such as `http://example.org/example?test test test`.