Change the auto-queue to run every hour
This will help minimize the number of scans in queue at any given time. It used to be that > 100 scans could be in the queue because they were all scheduled at one time. When a scan is queued, it also prevents a developer from running their own scan at a higher priority.